feat: rewrite date_utils with chrono, move business logic to shared crate
- Add chrono 0.4.44 (alloc feature, WASM-safe) to kartoteka-shared
- Create crates/shared/src/date_utils.rs with parse_date, format_date,
  add_days, day_of_week (Sun=0 JS convention), days_between, days_in_month,
  week_range, month_grid_range, prev_month, next_month, is_overdue,
  is_overdue_for_date_type, is_upcoming, sort_by_deadline (16 passing tests)
- Rewrite crates/frontend/src/components/common/date_utils.rs:
  re-export shared functions, keep presentation logic (Polish strings,
  js_sys wrappers), add day_of_week wrapper (Mon=0 convention for calendar),
  keep is_overdue/is_upcoming as 2-arg wrappers using current_time_hhmm()
- Update parse_date callers (calendar_nav.rs, calendar/mod.rs) to use
  NaiveDate methods (.year(), .month()) instead of tuple destructuring
- Add chrono dep to frontend Cargo.toml for Datelike trait in callers
- Fix add_days call sites in month_grid.rs to use i64 (shared API)

+++ crates/frontend/src/components/calendar/month_grid.rs
@@ -21,12 +21,12 @@ pub fn MonthGrid(counts: Vec<DaySummary>, year: i32, month: u32, today: String) 
     // Build grid: find first Monday before month start
     let first_day = format!("{:04}-{:02}-01", year, month);
     let first_dow = day_of_week(&first_day);
-    let grid_start = add_days(&first_day, -(first_dow as i32));
+    let grid_start = add_days(&first_day, -(first_dow as i64));
 
-    let dim = days_in_month(year, month) as u32;
+    let dim = days_in_month(year, month);
     let last_day = format!("{:04}-{:02}-{:02}", year, month, dim);
     let last_dow = day_of_week(&last_day);
-    let grid_end = add_days(&last_day, (6 - last_dow) as i32);
+    let grid_end = add_days(&last_day, (6 - last_dow) as i64);
 
     // Generate all dates in the grid
     let mut dates: Vec<String> = Vec::new();
